Refactor initialization of hazelcast for clustered mesh instances Refactor test helpers for clustering junit tests Add some clustering junit tests
