+++ mdm/hibernate-core/src/main/java/com/gentics/mesh/database/cluster/HibClusterManager.java
@@ -1,8 +1,11 @@
 package com.gentics.mesh.database.cluster;
 
+import static com.gentics.mesh.MeshEnv.CONFIG_FOLDERNAME;
+
+import java.io.File;
+import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.util.List;
-import java.util.Optional;
 import java.util.stream.Collectors;
 
 import javax.inject.Inject;
@@ -14,9 +17,10 @@ import com.gentics.mesh.core.rest.admin.cluster.ClusterInstanceInfo;
 import com.gentics.mesh.core.rest.admin.cluster.ClusterStatusResponse;
 import com.gentics.mesh.distributed.coordinator.MasterElector;
 import com.gentics.mesh.distributed.coordinator.MeshMemberInfo;
-import com.gentics.mesh.etc.config.ClusterOptions;
 import com.gentics.mesh.etc.config.HibernateMeshOptions;
 import com.gentics.mesh.util.UUIDUtil;
+import com.hazelcast.config.Config;
+import com.hazelcast.config.XmlConfigBuilder;
 import com.hazelcast.core.Hazelcast;
 import com.hazelcast.core.HazelcastInstance;
 
@@ -36,34 +40,41 @@ import io.vertx.spi.cluster.hazelcast.HazelcastClusterManager;
 public class HibClusterManager implements ClusterManager {
 	private static final Logger log = LoggerFactory.getLogger(HibClusterManager.class);
 
-	private final ClusterOptions clusterOptions;
+	private final HibernateMeshOptions options;
 	private final Lazy<Database> db;
 	private HazelcastClusterManager clusterManager;
 	private Lazy<MasterElector> masterElector;
+	private HazelcastInstance hazelcastInstance;
 
 	@Override
 	public HazelcastInstance getHazelcast() {
-		if (clusterManager == null) {
-			return Hazelcast.getAllHazelcastInstances().stream().findFirst().orElse(null);
+		if (hazelcastInstance == null) {
+			Config config = null;
+			String hazelcastFilePath = new File(CONFIG_FOLDERNAME, "hazelcast.xml").getAbsolutePath();
+			try {
+				config = new XmlConfigBuilder(hazelcastFilePath).build();
+			} catch (FileNotFoundException e) {
+				config = new Config();
+			}
+			config.setInstanceName(options.getNodeName());
+			config.setClusterName(options.getClusterOptions().getClusterName());
+			config.setClassLoader(Thread.currentThread().getContextClassLoader());
+			hazelcastInstance = Hazelcast.getOrCreateHazelcastInstance(config);
 		}
-		return clusterManager.getHazelcastInstance();
+		return hazelcastInstance;
 	}
 
 	@Override
 	public io.vertx.core.spi.cluster.ClusterManager getVertxClusterManager() {
 		if (clusterManager == null) {
-			Optional<HazelcastInstance> hazelcastInstance = Hazelcast.getAllHazelcastInstances().stream().findFirst();
-			clusterManager = hazelcastInstance
-					// when using second level cache, we already have an hazelcast instance
-					.map(instance -> new HazelcastClusterManager(instance))
-					.orElseGet(HazelcastClusterManager::new);
+			clusterManager = new HazelcastClusterManager(getHazelcast());
 		}
 		return clusterManager;
 	}
 
 	@Inject
 	public HibClusterManager(HibernateMeshOptions options, Lazy<Database> db, Lazy<MasterElector> masterElector) {
-		this.clusterOptions = options.getClusterOptions();
+		this.options = options;
 		this.db = db;
 		this.masterElector = masterElector;
 	}
@@ -75,7 +86,7 @@ public class HibClusterManager implements ClusterManager {
 
 	@Override
 	public ClusterStatusResponse getClusterStatus() {
-		List<ClusterInstanceInfo> instances = clusterManager.getHazelcastInstance().getCluster().getMembers().stream().map(member -> {
+		List<ClusterInstanceInfo> instances = getHazelcast().getCluster().getMembers().stream().map(member -> {
 			ClusterInstanceInfo info = new ClusterInstanceInfo();
 			info.setAddress(member.getAddress().toString());
 			info.setUuid(UUIDUtil.toShortUuid(member.getUuid()));
