+++ tests/common/src/main/java/com/gentics/mesh/test/local/MeshLocalServer.java
@@ -349,7 +349,17 @@ public class MeshLocalServer extends TestWatcher implements MeshTestServer {
 		public boolean clusterMode() {
 			return clustering;
 		}
-		
+
+		@Override
+		public String clusterName() {
+			return null;
+		}
+
+		@Override
+		public int clusterInstances() {
+			return 0;
+		}
+
 		@Override
 		public AWSTestMode awsContainer() {
 			return AWSTestMode.NONE;
