+++ tests/context-api/src/main/java/com/gentics/mesh/test/MeshInstanceProvider.java
@@ -1,5 +1,7 @@
 package com.gentics.mesh.test;
 
+import java.util.List;
+
 import org.testcontainers.utility.ThrowingFunction;
 
 import com.gentics.mesh.core.db.Database;
@@ -74,11 +76,11 @@ public interface MeshInstanceProvider<T extends MeshOptions> extends MeshOptions
 	/**
 	 * Performs the fast storage cleanup, e.g. in place data drop with keeping the schemas, if available.
 	 * 
-	 * @param db the database to clean up
+	 * @param dbs list of databases to clean up
 	 * @return true if the fast cleanup is available and succeeded.
 	 * @throws Exception
 	 */
-	default boolean fastStorageCleanup(Database db) throws Exception {
+	default boolean fastStorageCleanup(List<Database> dbs) throws Exception {
 		return false;
 	}
 }
