+++ tests/tests-distributed-coordinator/pom.xml
@@ -15,5 +15,9 @@
 			<groupId>com.gentics.mesh</groupId>
 			<artifactId>mesh-distributed-coordinator</artifactId>
 		</dependency>
+		<dependency>
+			<groupId>com.gentics.mesh</groupId>
+			<artifactId>mesh-test-common</artifactId>
+		</dependency>
 	</dependencies>
 </project>
