+++ verticles/graphql/src/main/java/com/gentics/mesh/graphql/filter/ImageDataFieldFilter.java
@@ -26,11 +26,11 @@ public abstract class ImageDataFieldFilter<E extends HibImageDataElement, T exte
 		List<FilterField<T, ?>> filters = super.getFilters();
 		//filters.add(makeWrappedFieldFilter("filename", "Filters by filename", StringFilter.filter(), HibImageDataField::getFileName));
 		//filters.add(makeWrappedFieldFilter("mime", "Filters by MIME type", StringFilter.filter(), HibImageDataField::getMimeType));
-		filters.add(makeWrappedFieldFilter("imageDominantColor", "Filters by image dominant color", StringFilter.filter(), HibImageDataField::getImageDominantColor));
-		filters.add(makeWrappedFieldFilter("plainText", "Filters by text data", StringFilter.filter(), HibImageDataField::getPlainText));
-		filters.add(makeWrappedFieldFilter("altitude", "Filters by altitude", NumberFilter.filter(), HibImageDataField::getLocationAltitude));
-		filters.add(makeWrappedFieldFilter("latitude", "Filters by latitude", NumberFilter.filter(), HibImageDataField::getLocationLatitude));
-		filters.add(makeWrappedFieldFilter("longitude", "Filters by longitude", NumberFilter.filter(), HibImageDataField::getLocationLongitude));	
+		filters.add(makeWrappedFieldFilter("imageDominantColor", "Filters by image dominant color", StringFilter.filter(), edge -> edge.getImageDominantColor()));
+		filters.add(makeWrappedFieldFilter("plainText", "Filters by text data", StringFilter.filter(), edge -> edge.getPlainText()));
+		filters.add(makeWrappedFieldFilter("altitude", "Filters by altitude", NumberFilter.filter(), edge -> edge.getLocationAltitude()));
+		filters.add(makeWrappedFieldFilter("latitude", "Filters by latitude", NumberFilter.filter(), edge -> edge.getLocationLatitude()));
+		filters.add(makeWrappedFieldFilter("longitude", "Filters by longitude", NumberFilter.filter(), edge -> edge.getLocationLongitude()));	
 		filters.add(makeWrappedFieldFilter("focalPointX", "Filters by longitude", NumberFilter.filter(), 
 				edge -> (edge != null && edge.getImageFocalPoint() != null) ? edge.getImageFocalPoint().getX() : null));	
 		filters.add(makeWrappedFieldFilter("focalPointY", "Filters by longitude", NumberFilter.filter(),
